Moved the audio import and queues to gamelayout
