+++ resources/js/Layouts/GameLayout.vue
@@ -1,6 +1,50 @@
 <script setup>
 import ApplicationLogo from '@/Components/ApplicationLogo.vue';
+import { ref } from "vue";
 import { Link } from '@inertiajs/vue3';
+import background from '@/../../public/audio/background.mp3';
+import correct from '@/../../public/audio/correct.wav';
+import incorrect from '@/../../public/audio/incorrect.wav';
+import chatChatter from '@/../../public/audio/catChatter.wav';
+import chatPurr from '@/../../public/audio/catPurr.wav';
+
+const backgroundAudio = new Audio(background);
+const correctAudio = new Audio(correct);
+const incorrectAudio = new Audio(incorrect);
+const catChatterAudio = new Audio(chatChatter);
+const catPurrAudio = new Audio(chatPurr);
+
+const levelCompleted = ref(false);
+
+const playBackground = () => {
+    if (levelCompleted.value === false) {
+        backgroundAudio.volume = 0.1;
+        backgroundAudio.play();
+    }
+}
+
+const levelComplete = () => {
+    correctAudio.play();
+    backgroundAudio.pause();
+    catPurrAudio.volume = 0.1
+    catPurrAudio.play();
+    levelCompleted.value = true;
+}
+
+const incorrectAnswer = () => {
+    incorrectAudio.play();
+}
+
+const catChatter = () => {
+    catChatterAudio.play();
+}
+
+defineExpose({
+    playBackground,
+    levelComplete,
+    incorrectAnswer,
+    catChatter,
+})
 </script>
 
 <template>
